requirements.txt
Django>=1.2.3
django-uni-form>=0.8.0
pyjwt>=1.5.0